+++ src/hotspot/share/jeandle/jeandleIntrinsicLowering.hpp
@@ -21,10 +21,12 @@
 #define SHARE_JEANDLE_INTRINSIC_LOWERING_HPP
 
 #include "jeandle/__llvmHeadersBegin__.hpp"
+#include "llvm/ADT/SmallVector.h"
+#include "llvm/IR/IRBuilder.h"
+#include "llvm/IR/InlineAsm.h"
+#include "llvm/IR/InstrTypes.h"
 #include "llvm/IR/Instruction.h"
 #include "llvm/IR/Instructions.h"
-
-#include "jeandle/jeandleCallSiteAttr.hpp"
 
 #include "jeandle/__hotspotHeadersBegin__.hpp"
 #include "classfile/vmIntrinsics.hpp"
@@ -33,6 +35,86 @@
 
 class JeandleAbstractInterpreter;
 class ciMethod;
+
+// =============================================================================
+// Control-flow facts for a lowered intrinsic call site. Combined into
+// CallSiteAttributeMetadata::control_flags with bitwise OR.
+// =============================================================================
+enum JeandleControlFlag : uint8_t {
+  CTRL_NONE                 = 0,
+  // The lowering can transfer control to uncommon_trap / deopt.
+  CTRL_MAY_DEOPT            = 1u << 0,
+  // The call may throw a Java exception and needs invoke-style exception
+  // continuation handling.
+  CTRL_NEEDS_EXCEPTION_EDGE = 1u << 1,
+};
+
+// =============================================================================
+// Memory-effect facts for a lowered intrinsic call site. Combined into
+// CallSiteAttributeMetadata::memory_flags with bitwise OR and translated into
+// LLVM call-site memory attributes where safe.
+// =============================================================================
+enum JeandleMemoryFlag : uint16_t {
+  MEM_NONE              = 0,
+  MEM_READ              = 1u << 0,
+  MEM_WRITE             = 1u << 1,
+  MEM_ORDERING_ONLY     = 1u << 2,
+  MEM_NEEDS_GC_STATE    = 1u << 3,
+};
+
+// =============================================================================
+// CallSiteAttributeMetadata — IR-level call-site facts. Narrowly focused on
+// attributes that affect LLVM IR lowering (deopt bundles, exception edges,
+// GC visibility, memory effects).
+// =============================================================================
+struct CallSiteAttributeMetadata {
+  uint8_t  control_flags;   // bitmask of JeandleControlFlag
+  uint16_t memory_flags;    // bitmask of JeandleMemoryFlag
+
+  bool may_deopt()            const { return (control_flags & CTRL_MAY_DEOPT) != 0; }
+  bool needs_exception_edge() const { return (control_flags & CTRL_NEEDS_EXCEPTION_EDGE) != 0; }
+  bool reads_memory()         const { return (memory_flags  & MEM_READ) != 0; }
+  bool writes_memory()        const { return (memory_flags  & MEM_WRITE) != 0; }
+  bool only_orders_memory()   const { return (memory_flags  & MEM_ORDERING_ONLY) != 0; }
+  bool needs_gc_state()       const { return (memory_flags  & MEM_NEEDS_GC_STATE) != 0; }
+  bool attach_deopt_bundle()  const {
+    return may_deopt() || needs_gc_state() || needs_exception_edge();
+  }
+  bool gc_leaf_by_flags()     const {
+    return !needs_gc_state() && !may_deopt() && !needs_exception_edge();
+  }
+};
+
+// =============================================================================
+// JeandleRuntimeCalleeFn — function pointer type for runtime stub resolvers.
+// Each resolver materializes the stub/SharedRuntime FunctionCallee in the
+// given LLVM module.
+// =============================================================================
+using JeandleRuntimeCalleeFn = llvm::FunctionCallee (*)(llvm::Module&);
+
+// =============================================================================
+// Call-site IR annotation helpers.
+// =============================================================================
+
+// Stamp the gc-leaf-function attribute on a call site when the call-site
+// metadata or runtime entry indicates a leaf call.
+void annotate_call(llvm::CallBase* call,
+                   const CallSiteAttributeMetadata& attrs,
+                   bool is_gc_leaf_entry = false);
+
+// Translate memory flags into LLVM call-site memory attributes. Only applied
+// when the call is safe from LLVM's reordering perspective (no GC-state, no
+// deopt, no exception edge).
+void apply_memory_attr(llvm::CallBase* call,
+                       const CallSiteAttributeMetadata& attrs);
+
+// Emit a side-effecting inline-asm call and mark it gc-leaf. Used by the
+// blackhole sink.
+llvm::CallInst* emit_gc_leaf_inline_asm(llvm::IRBuilder<>& builder,
+                                        llvm::FunctionType* fn_ty,
+                                        llvm::StringRef asm_string,
+                                        llvm::StringRef constraints,
+                                        llvm::ArrayRef<llvm::Value*> args);
 
 using JeandleTrapReasonMask = uint32_t;
 static_assert(Deoptimization::Reason_LIMIT <= 32,
@@ -57,6 +139,11 @@ class JeandleIntrinsicLowering : public StackObj {
  private:
   JeandleAbstractInterpreter* _interp;
   const ciMethod* _target;
+
+  // Arch-specific CPU feature checks. Defined in cpu/<arch>/jeandleIntrinsicLowering_<arch>.cpp.
+  static bool cpu_supports_rounding();   // floor/ceil/rint
+  static bool cpu_supports_popcount();   // bitCount_i/bitCount_l
+  static bool cpu_supports_spin_wait();  // onSpinWait
 
   // ========================================================================
   // Shared emit helpers
@@ -107,15 +194,5 @@ class JeandleIntrinsicLowering : public StackObj {
   // Attach JavaKlass/JavaKlassExact return-type attributes to a call site.
   void attach_callee_return_klass_attr(llvm::CallBase* call) const;
 };
-
-// =============================================================================
-// Arch-specific CPU feature checks for intrinsics.
-// Defined in cpu/<arch>/jeandleIntrinsicLowering_<arch>.cpp.
-// Return true if the current CPU supports the given intrinsic on this
-// architecture.
-// =============================================================================
-bool cpu_supports_rounding();   // floor/ceil/rint
-bool cpu_supports_popcount();   // bitCount_i/bitCount_l
-bool cpu_supports_spin_wait();  // onSpinWait
 
 #endif // SHARE_JEANDLE_INTRINSIC_LOWERING_HPP
